feat: Simpler Plugin System
This commit makes the connector apply the plugin as soon as the
plugin is defined. Assuming that the user awaited until the connector
was ready, it should always apply all plugins to all wallets.

Furthermore, this way allows the plugins to take the wallets and
decide what to do with them (if they want to intercept their
methods, or if they want to add new methods)